LLM APP/requirements.txt
groq==0.4.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
prophet==1.1.4
PyPDF2==3.0.1
python-docx==1.0.1
python-pptx==0.6.22
transformers==4.35.2
torch
faiss-cpu==1.7.4
Flask==3.0.0
Werkzeug==3.0.1
sentence-transformers==2.2.2
openpyxl==3.1.2